some major :poop: TDD right here
